feat(plugin-feed): fetch article on showItem; refine PostArticle chrome
Three changes:

- Extract the HTTP fetch + extract logic from the FetchArticleContent
  handler into a shared `util/fetch-article.ts` helper. The handler now
  wraps it via Effect.tryPromise. Same timeout, Content-Length cap,
  streamed byte cap, and hostname denylist as before; now reusable from
  the UI.
- MagazineArticle.handleOpen fires the fetch in the background (via
  fetchArticle) when a tile is opened via showItem. If the article has
  at least one image URL, writes the first one to post.imageUrl so the
  tile/PostArticle display the article's hero rather than the image
  embedded in the feed's description. Failures are logged and
  non-fatal — showItem still runs. readAt is set before the fetch is
  kicked off.
- PostArticle: title moved out of the toolbar into the article body as
  an h1. Toolbar now holds only an IconButton (arrow-square-out) that
  opens the original article in a new tab via window.open with
  noopener,noreferrer.
