+++ packages/plugins/plugin-feed/src/containers/PostArticle/PostArticle.tsx
@@ -2,10 +2,10 @@
 // Copyright 2026 DXOS.org
 //
 
-import React from 'react';
+import React, { useCallback } from 'react';
 
 import { type AppSurface } from '@dxos/app-toolkit/ui';
-import { Button, Panel, Toolbar, useTranslation } from '@dxos/react-ui';
+import { Panel, Toolbar, useTranslation } from '@dxos/react-ui';
 
 import { meta } from '#meta';
 import { type Subscription } from '#types';
@@ -18,23 +18,29 @@ export const PostArticle = ({ role, subject: post }: PostArticleProps) => {
   const { t } = useTranslation(meta.id);
   const metaLine = [post.author, formatDate(post.published)].filter(Boolean).join(' · ');
 
+  const handleOpenOriginal = useCallback(() => {
+    if (post.link) {
+      window.open(post.link, '_blank', 'noopener,noreferrer');
+    }
+  }, [post.link]);
+
   return (
     <Panel.Root role={role}>
       <Panel.Toolbar asChild>
         <Toolbar.Root>
-          <Toolbar.Text>{post.title ?? t('post-title.placeholder')}</Toolbar.Text>
-          <Toolbar.Separator />
           {post.link && (
-            <Button asChild variant='ghost'>
-              <a href={post.link} target='_blank' rel='noopener noreferrer'>
-                {t('open-original.label')}
-              </a>
-            </Button>
+            <Toolbar.IconButton
+              label={t('open-original.label')}
+              icon='ph--arrow-square-out--regular'
+              iconOnly
+              onClick={handleOpenOriginal}
+            />
           )}
         </Toolbar.Root>
       </Panel.Toolbar>
       <Panel.Content>
         <article className='flex flex-col gap-3 p-4 overflow-y-auto h-full'>
+          <h1 className='text-xl font-semibold'>{post.title ?? t('post-title.placeholder')}</h1>
           {metaLine && <div className='text-xs text-subdued'>{metaLine}</div>}
           {post.imageUrl && <img src={post.imageUrl} alt='' className='rounded w-full object-cover max-h-72' />}
           {post.snippet && <p className='text-sm whitespace-pre-wrap'>{post.snippet}</p>}
